BungalovChat/requirements.txt
sentence-transformers
fastapi
uvicorn
pydantic
